Change /cooperatives path to /business path.
