+++ auth-api/email_templates/staff_review_account_email.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Notification from Cooperatives Online</title>
+    <style>
+        /* Global Resets */
+        body,
+        table,
+        div,
+        a,
+        p {
+            font-family: Verdana, Arial, sans-serif;
+        }
+
+        body {
+            margin: 0;
+            padding: 0;
+        }
+
+        table {
+            border-collapse: collapse;
+        }
+
+        img {
+            border: none;
+        }
+    </style>
+    <meta name="referrer" content="origin-when-cross-origin" />
+    <title>chat.pathfinder.gov.bc.ca</title>
+    <meta name="application-name" content="chat.pathfinder.gov.bc.ca">
+    <meta name="apple-mobile-web-app-title" content="chat.pathfinder.gov.bc.ca">
+    <meta http-equiv="content-language" content="">
+    <meta name="language" content="">
+    <base href="/">
+</head>
+
+<body>
+    <table style="width: 100%;" cellspacing="0" cellpadding="0">
+        <tr>
+            <td style="padding-left: 20px; padding-right: 20px;">
+                <table
+                    style="max-width: 640px; margin: 0 auto; border: 0; border-collapse: collapse; border-radius: 6px; background-color: #ffffff;"
+                    cellspacing="0" cellpadding="0">
+                    <tr>
+                        <td style="padding-top: 30px; padding-bottom: 30px;">
+                            <img src="{{ logo_url }}" height="54" alt="Government of British Columbia" />
+                        </td>
+                    </tr>
+                    <tr>
+                        <td>
+                            <h1
+                                style="margin: 0; line-height: 34px; letter-spacing: -1px; font-family: Verdana, Arial, sans-serif; font-size: 24px; font-weight: 700;">
+                                <strong>{{ user.firstname }} {{ user.lastname }}</strong> has submitted a request that
+                                requires your approval. You will need to log in using your IDIR to approve this request.
+                            </h1>
+                            <p
+                                style="margin-top: 20px; margin-bottom: 0; line-height: 24px; font-family: Verdana, Arial, sans-serif; font-size: 16px;">
+                                Log into Cooperatives Online to approve this .
+                            </p>
+                            <div style="margin-top: 45px;">
+                                <!--[if mso]> <v:roundrect xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="urn:schemas-microsoft-com:office:word" href="{{ url }}" style="height:55px; v-text-anchor:middle; width:250px;" arcsize="10%" stroke="f" fillcolor="#003366" alias="CTA1">Join Team<w:anchorlock/> <center> <![endif]-->
+                                <a style="display: block; width: 250px; padding: 15px 0; color: #ffffff; border: none; border-radius: 6px; background-color: #003366; text-align: center; text-decoration: none; font-size: 16px; font-weight: bold;"
+                                    href="{{ url }}">Cooperatives Online
+                                </a>
+                                <!--[if mso]> </center> </v:roundrect> <![endif]-->
+                            </div>
+                        </td>
+                    </tr>
+                </table>
+            </td>
+        </tr>
+    </table>
+</body>
+
+</html>
