+++ auth-web/.env.production
@@ -1,2 +1,2 @@
-VUE_APP_PATH = /cooperatives/auth/
-VUE_APP_PATH_COOPS = cooperatives
+VUE_APP_PATH = /business/auth/
+VUE_APP_PATH_COOPS = business
