+++ auth-web/nginx.conf
@@ -19,26 +19,26 @@ http {
  default_type  application/octet-stream;
  server_tokens off;
  underscores_in_headers on;
- 
+
  # Use a w3c standard log format
  log_format  main  '$remote_addr - $remote_user [$time_local] "$request" '
                       '$status $body_bytes_sent "$http_referer" '
                       '"$http_user_agent" "$http_x_forwarded_for"';
- 
+
  access_log  /var/log/nginx/access.log  main;
 
- 
+
  server {
-  
+
    # add in most common security headers
    add_header Content-Security-Policy "default-src * data: blob: filesystem: 'unsafe-inline' 'unsafe-eval'";
    add_header Strict-Transport-Security "max-age=31536000; includeSubDomains";
    add_header X-Content-Type-Options "nosniff";
    add_header X-XSS-Protection 1;
    add_header X-Frame-Options DENY;
-  
-  
-  
+
+
+
    listen 8080;
    server_name _;
 
@@ -47,7 +47,7 @@ http {
    access_log /dev/stdout;
    root /app;
 
-   location /cooperatives/auth {
+   location /business/auth {
      alias /app;
      try_files $uri /index.html;
      gzip on;
