--- auth-web/openshift/Caddyfile
@@ -1,13 +1,0 @@
-0.0.0.0:2015
-
-root /var/www/html
-
-log stdout
-
-errors stdout
-
-rewrite /auth {
-    regexp .*
-    to {path} /cooperatives/auth/
-}
-
