+++ auth-web/openshift/templates/auth-web-pre-deploy.json
@@ -78,7 +78,7 @@
       "displayName": "WEB_APP_CONTEXT_PATH",
       "description": "The path at which web application is deployed.Context root for the web applicaton",
       "required": true,
-      "value": "cooperatives/auth"
+      "value": "business/auth"
     }
   ]
 }
