+++ auth-web/public/config/configuration.json
@@ -1,5 +1,5 @@
 {
-  "AUTH_URL": "https://dev.bcregistry.ca/cooperatives/auth/",
+  "AUTH_URL": "https://dev.bcregistry.ca/business/auth/",
   "VUE_APP_PAY_ROOT_API": "https://pay-api-dev.pathfinder.gov.bc.ca/api/v1",
   "VUE_APP_AUTH_ROOT_API": "https://auth-api-dev.pathfinder.gov.bc.ca/api/v1",
   "VUE_APP_LEGAL_ROOT_API": "https://legal-api-dev.pathfinder.gov.bc.ca/api/v1",
