+++ codecov.yaml
@@ -22,12 +22,6 @@ coverage:
         target: 35%
         flags:
           - authweb
-    branches:
-      - development
-    if_not_found: success
-    if_ci_failed: error
-    informational: false
-    only_pulls: true
 
 ignore:
   - "^/tests/**/*" # ignore test harness code
